# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chronus Booking API",
    "version": "1.0.0",
    "description": "API documentation for the Chronus multi-tenant mentor booking application."
  },
  "servers": [
    {
      "url": "http://localhost:3010/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/mentors": {
      "get": {
        "summary": "List Organization Mentors",
        "description": "Fetches active mentors in the authenticated user's organization with pagination support. Requires the session cookie.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number (1-indexed, default: 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of mentors per page (default: 10, max: 100)",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of mentors in the organization",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "membershipId": { "type": "string", "format": "uuid" },
                          "userId": { "type": "string", "format": "uuid" },
                          "email": { "type": "string", "format": "email" },
                          "name": { "type": "string" },
                          "timezone": { "type": "string", "example": "America/New_York" },
                          "createdAt": { "type": "string", "format": "date-time" },
                          "updatedAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer", "example": 25 },
                        "page": { "type": "integer", "example": 1 },
                        "limit": { "type": "integer", "example": 10 },
                        "totalPages": { "type": "integer", "example": 3 }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid session"
          },
          "403": {
            "description": "Membership deactivated or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/{orgId}/mentors": {
      "get": {
        "summary": "List Organization Mentors by orgId parameter",
        "description": "Fetches active mentors by orgId parameter, enforcing that the orgId matches the authenticated user's organizationId claims.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "description": "The Organization ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number (1-indexed, default: 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of mentors per page (default: 10, max: 100)",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of mentors in the organization",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "membershipId": { "type": "string", "format": "uuid" },
                          "userId": { "type": "string", "format": "uuid" },
                          "email": { "type": "string", "format": "email" },
                          "name": { "type": "string" },
                          "timezone": { "type": "string", "example": "America/New_York" },
                          "createdAt": { "type": "string", "format": "date-time" },
                          "updatedAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer", "example": 25 },
                        "page": { "type": "integer", "example": 1 },
                        "limit": { "type": "integer", "example": 10 },
                        "totalPages": { "type": "integer", "example": 3 }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid organization ID format"
          },
          "401": {
            "description": "Authentication required or invalid session"
          },
          "403": {
            "description": "Access denied due to tenant boundary validation mismatch"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/mentors/{mentorId}/slots": {
      "get": {
        "summary": "List Mentor Slots",
        "description": "Fetches availability slots for a specific mentor within the authenticated user's organization. Supports date range filtering via startDate and endDate. Requires the session cookie.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mentorId",
            "in": "path",
            "required": true,
            "description": "The Mentor's membership ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "description": "Filter slots starting at or after this ISO 8601 timestamp",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "description": "Filter slots starting at or before this ISO 8601 timestamp (defaults to +30 days from startDate)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of slots for the mentor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "organizationId": { "type": "string", "format": "uuid" },
                      "mentorId": { "type": "string", "format": "uuid" },
                      "startTime": { "type": "string", "format": "date-time" },
                      "endTime": { "type": "string", "format": "date-time" },
                      "status": { "type": "string", "enum": ["AVAILABLE", "BOOKED"] },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "updatedAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid mentor ID or date range format"
          },
          "401": {
            "description": "Authentication required or invalid session"
          },
          "403": {
            "description": "Access denied due to invalid membership"
          },
          "404": {
            "description": "Mentor not found in organization"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/{orgId}/mentors/{mentorId}/slots": {
      "get": {
        "summary": "List Mentor Slots by orgId parameter",
        "description": "Fetches availability slots for a specific mentor by orgId parameter, enforcing that the orgId matches the authenticated user's organizationId claims.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "description": "The Organization ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "mentorId",
            "in": "path",
            "required": true,
            "description": "The Mentor's membership ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "description": "Filter slots starting at or after this ISO 8601 timestamp",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "description": "Filter slots starting at or before this ISO 8601 timestamp (defaults to +30 days from startDate)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of slots for the mentor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "organizationId": { "type": "string", "format": "uuid" },
                      "mentorId": { "type": "string", "format": "uuid" },
                      "startTime": { "type": "string", "format": "date-time" },
                      "endTime": { "type": "string", "format": "date-time" },
                      "status": { "type": "string", "enum": ["AVAILABLE", "BOOKED"] },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "updatedAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter format"
          },
          "401": {
            "description": "Authentication required or invalid session"
          },
          "403": {
            "description": "Access denied due to tenant boundary validation mismatch or invalid membership"
          },
          "404": {
            "description": "Mentor not found in organization"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  }
}